split out test_schema
